Rename test file and add a describe block
